u-boot-iot2050: Add recipe for building upstream queue
This is still work-in-progress, so the default remains on the TI-SDK
version. However, the selected queue is already able to run on both
variants, just without signature checking at the time of writing. We
reference the branch, not a specific revision, so that always the latest
version is pulled until everything is merged, and we can freeze specific
versions.

The upstream recipe uses the new, separate atf and op-tee packages. It
also comes with a new version of SE-Boot, unified for both variants,
only working with the upstream U-Boot.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>