npm: Split buildchroot from target image usage, update the latter
This is currently the best we can do: Latest npm 6.14.x as Debian
backports now provide gets stuck on npm pack during the fetch step.
This is apparently fixed in 7.0.x as bullseye will provide, but
backporting that is too much effort and brings too many changes to the
target image.

So rename the npm package used for fetching and installing to
"npm-buildchroot", keeping the currently used version untouched. For the
target, move to Debian buster-backports so that we benefit from security
updates for npm and its node-* dependencies.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>