npm: Add support for using prebuilt package again
Debian just updated npm in buster-backports, just
https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=980928 needs to be
fixed, and then we will be fine with the binary package. Also Ubuntu is
delivering the right one already.

Give the distro control over that. It can set NPM_CLASS_PACKAGE to "npm"
again and switch OWN_NPM_CLASS_PACKAGE to 0 so that no recipe is
expected for npm.

For now, nothing changes for buster.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>