u-boot: Add NOR protection support for Winbond chips
IOT2050 use a Winbond chip, and it supports lock/unlock function,
so that user can use `sf protect lock` and `sf protect unlock` commands
to lock/unlock the protection, so that the content of the flash could
be protected whenever needed.

This commit is to enable the NOR protect support for Winbond chips.

Signed-off-by: Su Baocheng <baocheng.su@siemens.com>