npm: Switch to version from bullseye
This is not as simple as it appeared: Even with the reworked local
bundling of fetched packages and deps, the npm 5.x version from buster
does not work for us. The version currently in buster-backports is
broken, via multiple deps. So we need to switch to bullseye.

With bullseye, we have to be careful not picking packages that have
strict dependencies on arch-specific packages. That is fortunately only
the case for node-typedarray-to-buffer (still taken from buster) and
node-psl (for which buster-backports remain needed).

Besides node-*, we also need to pick gyp from bullseye. That brings the
change that this package pulls python3 and does not establish
/usr/bin/python - which is in buster still python2. This python link is
needed by some nodejs package scripts, though. We mitigate that by also
installing python as fetch dependencies of our npm recipes.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>