mraa: Update to 2.2.0 and new GPIO line location scheme
This brings a number of updates to mraa. Specifically, it switches to
using GPIO line names for discovering their sysfs numbers. This is
needed to support upstream kernels now and systems with different
gpiochip probing orders in the future.

Furthermore, it adds support the the user button.

By moving to 2.2.0, we now need to explicitly enable javascript support
during configure.

Fixes #23
Fixes #24
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>