Add support for interactive and automatic eMMC installation
Add an early systemd unit that writes the so far unmodified and
read-only mounted SD card (or USB stick) to the eMMC of the Advanced
device. This is either triggered by pressing and holder the USER button
for 5 seconds during the very first boot, or by previously creating the
file /etc/install-on-emmc on the image.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>