openssl: Pin back to buster release
"apt-get source" is not respecting apt-preferences. And since we have
both buster and bullseye configured as sources, we always get the latest
version - from bullseye. This led to 3ecb19a1a1ee which was wrong.

Switch back to buster's original openssl version by selecting the
release in the SRC_URI - which luckily "just works". We are using
BASE_DISTRO_CODENAME to ease the support of Ubuntu 20.04 via the same
recipe.

And then revert 3ecb19a1a1ee.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>