Switch to Debian 11 (bullseye)
This requires additional changes to mraa, namely the removal of the
swig3.0 hack, replacing it with the self-built swig4.0 packages and two
additional fixes for mraa.

The distro config can luckily be freed from a lot of backports. The
machine config was still carrying HOST_DISTRO, although that was no
longer required by Isar for a longer while.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>