install-on-emmc: Fix partition table update after image writing
partx -a is recommended to inform the kernel about new partitions now
visible when rescanning the device. But that seems to be only one piece.
It may take some time for the device node to show up. So wait until that
happened. This should fix errors like

install-on-emmc.sh[175]: Updating partition UUID of eMMC rootfs
install-on-emmc.sh[175]: partx: specified range <1:0> does not make sense
install-on-emmc.sh[175]: mount: /mnt: special device /dev/mmcblk1p1 does not exist.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>