iot2050-firmware-update: Remove checksumming
Checksumming could be easily done by user, so de-scope the checksumming
from the tool.
So now the force mode will be defined to force to update a binary to
flash regardless of all the pre-defined limitations, such as soc ID, or
some new limitations that will be added in the future.

Signed-off-by: Su Bao Cheng <baocheng.su@siemens.com>