iot2050:u-boot:change the boot targets
For our advanced board,our first boot target is emmc and there is industry
os resides in it. And in the industry os, there is no method to change the
boot order. So the user can not boot from media by default.

so we need to change the boot target as usb --> uSD --> emmc.
Make it possible for user to boot the image from other media.

Signed-off-by: Chao Zeng <chao.zeng@siemens.com>