customizations-example: Resolve renaming race when swapping eth0/eth1
Using eth* as target name is racy against the kernel naming its
interfaces this way in parallel. Somehow, Debian 10 (or the udev version
therein) managed to rename the interface to a temporary name first and
only afterwards to the final one. This is no longer working with Debian
11 and also Ubuntu 20.04.

Fix this by assigning names from the Consistent Network Device Naming
scheme: X1-P1 (eth0) becomes eno1, X1-P2 (eth1) eno2. We still rely on
the probing ordering of eth0/eth1 not changing. It's the only generic
anchor we have because Ethernet MACs are not predictable.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>