u-boot-iot2050: Update patches to support PG2 and drop network boot
This aligns the patch queue for 2021.04 with the series targeting
upstream, adding support for Product Generation 2 (PG2), aligning the
device trees with the kernel queue and dropping incomplete Ethernet boot
support via prueth. The latter can be added only once TI finally
upstreams their driver.

As Ethernet is removed, we can also remove the dependency on
ti-pruss-firmware. Also to be added again once the related feature is
ready.

The RTI watchdog firmware is now a common dependency of upstream and
2021.04 stable, so move it over to the include.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>