iot2050-image: Warn users about firmware incompatibility
By now, many users are familiar with building own imaging to get latest
features and fixes. As we are now requiring a firmware update for
supporting the 5.10 kernel, reports arrive that things are unexpectedly
broken.

Warn about this situation prominently during the image build. This
warning can later be adjusted to "requires firmware release ..." or
completely dropped again.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>