iot2050-image-swu-example: Add watchdog
Add function to start the watchdog before selecting a rootfs to boot.
The watchdog ensures a bootable system in case a update fails.

The systemd 'RuntimeWatchdogSec' has a fixed value of 60s
as the watchdog driver doesn't support updating the timeout
after start. Changing this value will trigger periodic reboots.

Signed-off-by: Quirin Gylstorff <quirin.gylstorff@siemens.com>