Update tiboot3.bin to V01.02.01.02
This tiboot3.bin fixes the wrong order of the MAC address according to
the printed order on the housing.

Please be noted, if the V01.01.01 example image is used, update to this
version of firmware will cause the MAC addresses reversed, if the MAC
address is used, for example, for the DHCP MAC binding/filtering, be
careful before updating to this firmware.

Signed-off-by: Su Bao Cheng <baocheng.su@siemens.com>