linux-iot2050:fix: couldn't connect to phy
When booting up,it would show the error "couldn't connect to phy".
as it trigger the deferred probe mechanism,it would probe later

But there is a chance of error, causing the MAC to fail to link with
the phy normally.Resulting in the ethernet can not work

How to reproduce:
1.power on and get the ip then ping router
2.reboot
3.waiting bootup,and repeat 1~3

Repeat about 500 times,it would cause the error.
Compile the mdio into kernel to fix this issue.

Signed-off-by: chao zeng <chao.zeng@siemens.com>