install-on-emmc: Add a case where the button has been pressed
As we can press the button to remove the emmc from boot_targets,
besides they'd like to install image to emmc. It is hard for user
to press the button,release the button then again pressing the button
to trigger the install process.

The easiest way is to keep pressing the button without thinking
So add one condition to detect the key has alreay been pressed always.

Signed-off-by: chao zeng <chao.zeng@siemens.com>