linux-iot2050: Fix the 8250 RS485 issue
An issue is observed on 8250 tty that under the RS485 or RS422 mode, the
RX/TX does not work. For RS485 the RX does not work meanwhile for RS422
the TX does not work.

This issue was introduced in the kernel driver when Lukas Wunner added
commit f45709df7731 to prevent the userspace interfering w/ RS485.
However this commit breaks the RS485.

Signed-off-by: Su Bao Cheng <baocheng.su@siemens.com>