Add Google Coral support via meta-coral
This uses the Isar layer meta-coral to build the required driver and
userland packages from source, including them by default in our images.
As the build time increased significantly by this (about 30 min. at
least), make the packages opt-in via a kas option and a menu entry.

As this addition requires pulling pending deb-profile patches which
change the recipe API, we need to adjust the openssl recipe to that as
well.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>