Update to latest Isar revision
This allows to drop one patch but currently requires a new one to fix
cross-build issues. Furthermore, we need to adjust to IMAGE_FSTYPES and
fix the swig dependency (swig is now built as :all).

Due to the introduction of sstate support, the input that
u-boot-iot2050-pg1/2:do_deploy uses has to be cached and uncached
explicitly so that partial rebuilds works.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>