optee-os: Update to 3.16
For us, this primarily provides fixes at this point. One exception:
We now get a warning that "This OP-TEE configuration might be insecure!"
since OP-TEE commit
https://github.com/OP-TEE/optee_os/commit/9e42008d9512dda216db2235576b8956e95c408c.
This is at least true as long as we do have support for a hardware
unique key, so leave this warning untouched for now.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>