Update isar-cip-core layer
Brings internal simplifications and allows to update to kas 3.x, thus
Debian bullseye as build environment. It also permits to build EFI Boot
Guard for arm64 which will be used soon.

Temporarily, we pick up an older version of wic-swu-img.bbclass in order
to preserve the r/w property of the rootfs that isar-cip-core already
dropped. Alignment to isar-cip-core will happen in succeeding steps.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>