iot2050-image-swu-example: Switch to UEFI boot and read-only rootfs
Adopting the UEFI-based boot that isar-cip-core pre-integrates allows
sharing of much more recipes and patterns. It also enables the switch to
the more appropriate read-only rootfs setup which avoids mixing up local
installation with full-image updating. Finally, this lays the ground for
secure boot and rootfs image integrity checking in later steps.

We boot the kernel now via EFI Boot Guard and a unified kernel image
generated by it. That means, watchdog handling will remain with U-Boot
but rootfs selection is moved to EFI Boot Guard.

The unified kernel image contains all device trees of the IOT2050
variants and selects the fitting one based on what U-Boot passed down as
firmware DTB. So we preserve the ability to update kernel and DTB in
lock-step.

An initramfs is now generate with pairs the booted kernel with the
associated rootfs and mounts the /etc overlay early.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>