install-on-emmc:use boot dev uuid as the emmc original uuid
When dd sd image to emmc, the uuid of the sd and emmc should keep
the same.But using some of sd cards to update the emmc,the uuid of
emmc is not the same with sd card. It would result in can not update
the uuid in the boot.scr.
next boot from emmc, it would hang at waitng for partuuid ...

So pass the boot dev partuuid as the original emmc partuuid,to avoid this bug.

Signed-off-by: chao zeng <chao.zeng@siemens.com>