regen-rootfs-uuid: accept parameter as UUID
Using the sfdisk can not get the real partuuid from the disk
if the disk gpt primary partition table corrupt sometimes.

Currently,found when dd the boot disk to emmc would result in
the gpt partition table corrupt in some cases.we can not get get
the right old uuid to replace new one in the boot script.

Extend this script to accept one parameter as the old uuid
in that case we can replace the partuuid in the boot.scr

Signed-off-by: chao zeng <chao.zeng@siemens.com>