Add recipe secure-boot-otp-provisioning
Add this package to generate the otpcmd binary for the customer key
provisioning.

This tool support both provisioning the key sets and switching the
current key to next key.

This package prepares for the upcoming integration into the u-boot
building, the generated artifacts of this package will be consumed by
u-boot building.

Signed-off-by: Baocheng Su <baocheng.su@siemens.com>