secure-boot: Re-organize the secure boot keys
The current key organization is kind of messy. Put all the keys into the
secure-boot-otp-provisioning, because this is by nature the first place
keys should be generated. Then make symbolics to this place from u-boot
and UEFI signing secrets.

Meanwhile format `file://dir` in SRC_URI does not handle symbolics very
well, if the symbolics are linked to files outside of the dir. So expand
it to cover such case.

Signed-off-by: Baocheng Su <baocheng.su@siemens.com>