iot2050-firmware-update: Optimization prompt display
Optimize the prompt process. No matter use '-f' force update or not,
it may render the device.

Currently, if user would like to keep the enviroment, they need to provide
u-boot-initial-env file of the target firmware.It can not be got from our release image.
So when using "-f" force update option, it would not preserve the environment now.
It would erase all the u-boot environment.

Signed-off-by: chao zeng <chao.zeng@siemens.com>