iot2050-firmware-update: Fix env that preserved is not right
Issue:
we used this tool on the previous release image found that if
there were two identical values will be used to sort higher.
However, the backward value is used on the latest image.

It is unreasonable to have two identical environment variables in
the same uboot env file. So remove the same environment variables
from the default list to fix this issue.

Signed-off-by: chao zeng <chao.zeng@siemens.com>