u-boot-iot2050: Rework support for M.2 connector configuration
This changes the interface between firmware and kernel for configuring
and communicating the M.2 mode.

Previously, a different device tree was selected by the firmware when a
different mode was set, including repeating updates of the U-Boot
environment in flash on every boot. This approach here uses the device
tree fix-up of U-Boot to adjust a given DT to the selected mode.

Further changes:
 - permit manual configuration in secure boot mode
 - report selected mode and if it was chosen manually or as fallback
 - align U-Boot DT to kernel DT
 - document manual configuration
 - cleanups and simplifications

Note: This temporarily breaks compatibility with the current kernel.
The kernel will be adjusted in the succeeding commit.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>