Integrate new tool k3conf
Introduction:
K3CONF is a Linux user-space standalone application designed to provide a quick'n easy way
to dynamically diagnose Texas Instruments' K3 architecture based processors
link:https://git.ti.com/cgit/k3conf/k3conf/about/

It's very useful tool to get and set the clock and so on.
For current usage, we can use this tool to switch cpu frequency from user-space.

Signed-off-by: chao zeng <chao.zeng@siemens.com>