customizations-example: Resizing a terminal
Unlike ssh, serial connections do not have a mechanism to transfer
something like SIGWINCH when a terminal is resized.
So it can not adjust the window size automatically.

This change would make adjust the console terminal column and row
according to the actual terminal window size when login or using
command line terminal-resize to resize a terminal.

Refer:https://wiki.osdev.org/Terminals

Signed-off-by: chao zeng <chao.zeng@siemens.com>