fix: auto-resolve native deps of npm modules
The patch adds the common shlibs:Depends pattern to the npm class.
This is required, as we use dpkg-raw as a base which does not add these
as default. By adding it, all native dependencies of a package providing
an NPM module are added automatically. This helps to avoid missing
runtime dependencies and in general makes the dependency graph more
precise.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>