iot2050-firmware-update: Add the rollback firmware feature
When the upgrade fails, update the previous firmware again.
Every effort is made to ensure that the equipment will not
be damaged.

At the same time, a command line interface has been added.
After the upgrade is completed, you can use this command to
roll back to a version before the upgrade.

Signed-off-by: chao zeng <chao.zeng@siemens.com>