feat: rework node-red directory layout
This patch changes the install location of the node-red modules.
All node-red-* modules are now installed in /usr/lib. By that,
we also do no longer need to create a NPM package for the preinstalled
nodes. Instead, we simply model this as a debian meta-package to
transitively install the "preinstalled" modules.

With this pattern, updates of the preinstalled modules via the palette
are still possible, as node-red installs the updates into the userDir.
The modules in userDir have a higher precedence.

This solves mainly two issues:
1. Move modules out of /root to later add support to update an image
   containing node-red using swupdate.
2. Prepare to limit the access rights of the node-red user, which also
   executes the webserver.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>