linux-iot2050: dts: fix the gpio configuration of m.2
For the nodes main_gpio0 and main_gpio1,
when we add the property pinctrl-0, it would override the
previous property not to add to the previous node.
So it would lack some configurations.

The configuration for the cp2102 and part of arduino pins
are not correct.
Add the property back to solve this issue.

Signed-off-by: chao zeng <chao.zeng@siemens.com>