patch-u-boot-env: add missing rt dep to config
The patch-u-boot-env package requires a u-boot-config file.
By that, we also need a dependency to a provider of the config.
Just relying on the ambient package provided by the BSP is error prone.

To break the bb dependency chain, we also ensure that all components
the u-boot bb recipe provides are also provided as runtime dependency.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>