allow to override initramfs conf of iot2050-debian
This patch allows to derive custom swupdate images from the
iot2050-debian distro. Previously, derived swu images could not be
created as the abrootfs image name could not be overwritten, hence
always pointed to the example image. This led to hard-to-debug
situations where the boot failed as the image UUIDs did not match.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>