Update layers to bitbake 2
Besides the infamous syntax conversion and marking sudo tasks as such,
this also brings some updates via the layers themselves.

Most notably for this commit are changes in Isar that prefill the
sbuild-chroot with crossbuild dependencies, also for the compat case.
That brought back problems with rebuilding openssl which should now be -
once and for all - resolved by Isar patches bringing multiarch build and
auto-dependencies on essential packages that are rebuilt.

On isar-cip-core side, initramfs-etc-overlay-hook was renamed and the
ABROOTFS_PART_UUID_* variables became mandatory.

Finally, Isar now supports us in rebuilding essential packages, and that
for all archs we need (target, compat and build host). This not only
ensures that the package is available at the right time for all of them,
it also simplifies the rebuild, allowing us to drop workarounds like
preferences.upstream-libssl.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>