Update isar-cip-core
Brings minor changes due to different swupdate image modeling. We need
to activate the new IMAGE_FSTYPES "swu" and specify SWU_ROOTFS_TYPE as
verity in the secure boot case. IMAGE_TYPEDEPs should be expanded, not
set, to align with upstream. And IMAGE_CLASSES are now auto-inherited
via the layer.conf of isar-cip-core.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>