Add helper for efivarfs manipulating
This package helps to prepare the operating environment for manipulating
the UEFI authenticated variable.

The UEFI authenticated variables are created as immutable files[1], this
helper could help to remove the immutable flag before using the efitools
to manipulate them.

Besides, it also checks if the efivarfs is correctly mounted, or if the
efitools package is pre-installed.

[1]. https://www.kernel.org/doc/html/v5.10/filesystems/efivarfs.html

Signed-off-by: Baocheng Su <baocheng.su@siemens.com>