Make m.2 power control also avaiable to minipcie
Issue is that sometimes, some miniPCIE card hangs, due to
firmware/software issue. Such hang could not be recovered by software
reset, since it is not power on reset.

However, the power supply design on M.2 variant, that control the 3.3V
power pin through a GPIO is proven to work against such scenario. With
this GPIO, during boot, a POR will be triggered to fully recover the
miniPCIE or M.2 module.

So port this design also to the miniPCIE variants.

Signed-off-by: Baocheng Su <baocheng.su@siemens.com>