add service to acquire hw information for swupdate
This patch adds a one-shot systemd service that reads the hardware
revision part from /proc/device-tree/compatible and writes that to
/etc/hwrevision. By that, the hardware-compatibility node in the
swupdate description file can be used to limit the boards an update can
be applied to. This helps to avoid that images are applied to
unsupported boards (e.g. due to lack of RAM, no M2, etc...).

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>