iot2050-initramfs: Add crypt and ftpm hook for disk encryption
Integrate the data/disk encryption implementation to the secure boot. The
implementation is based on the OPTee fTPM (with MS's reference fTPM) and the
clevis and LUKS.

The introduction of clevis is due to the absence of systemd-cryptsetup in Debian
Bullseye, that means it could replaced by later one once Debian Bookworm is
released and integrated onto IOT2050.

Signed-off-by: Baocheng Su <baocheng.su@siemens.com>