Move optee-client to isar-cip-core
Prepares for upstreaming it into isar-cip-core.

The optee-client is one of the fundamentals for security of ARM
embedded, so a shared common repo is a better place to live.

The patch upstreaming is on-going, once done, the reviewed upstream
patch should take the place of the current ones, or a newer
isar-cip-core version should be picked to include them.

Signed-off-by: Baocheng Su <baocheng.su@siemens.com>