iot2050-firmware-update: Refactor the firmware update tool
At the beginning, a additional firmware was intended to be integrated
into the upgrade tarball package. And the original code was kind of
process-oriented code, many common features are not extracted, many
duplicate codes exist, and the code coupling is very high, which is very
difficult to expand and maintain. It was necessary to refactor the code for
better maintainability and extension.

Anyway, in the end the additional firmware is planed to integrated into its
associated subsystem for decoupling. Refactored code is preserved for
possible  future maintainability and extension. For example, if new firmware
needs to be added in the future, whether it is the existing firmware upgrade
method or other upgrade methods, the previous upgrade logic or upgrade
method can be more easily introduced and reused.

Signed-off-by: Li Hua Qian <huaqian.li@siemens.com>