u-boot-iot2050: Lift to 2023.10
Most of our patches are merged now, just one DT change, a small fix for
the signing script and a fix for the board detection are remaining.

In this latest upstream version, we folded PG1 and PG2 U-Boot configs
into a single one again, generating both artifacts in the same run.
Consequently, we can simplify the build as there is now only one recipe
as well.

We need to drop the disabling of CONFIG_DISTRO_DEFAULTS from the secure
boot configuration as this would make distro_bootcmd unavailable.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>