linux-iot2050: Update to v6.1.54-cip6[-rt3]
Move PRUETH driver from module to built-in in order to avoid a probing
race condition.

CONFIG_FB is no longer auto-selected. Enable it to preserve framebuffer
support.

CONFIG_PWM needs to be explicitly enabled with the newer kernel.

Finally, CONFIG_RT_GROUP_SCHED must be disabled for the RT kernel as it
is officially incompatible and breaks requesting RT scheduling policies.

Signed-off-by: Diogo Ivo <diogo.ivo@siemens.com>
[Jan: add phys_port_name icssg-prueth patch, fix defconfig]
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>