linux-iot2050, mraa: Update kernel patches to latest upstream queue
This primarily updates our own DT patches after their reviews and
checks. That change also needs small adjustments of the MRAA path for
debugfs-based pinmuxing due to node renamings.

While updating, also include more of the DT cleanups that are in latest
upstream so that we are closer to the official DT.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>