u-boot-iot2050: Add the support for the SM variant
The dts for the new SM variant is idential to the kernel.

Since the introduction of the 4G DDR on SM variant, the memory size is
now determined by the SEBoot and then passed to the u-boot via scratchpad
board info.

Also the fdt fixup logic is generalized.

BTW a minor coding style change for the iot2050/board.c is hitchhiked.

Signed-off-by: Chao Zeng <chao.zeng@siemens.com>
Signed-off-by: Li Hua Qian <huaqian.li@siemens.com>
Signed-off-by: Baocheng Su <baocheng.su@siemens.com>