iot2050-eio-manager: Add module configuration
The module configuration is written in YAML, which will be validated via
JSON schemas internally.

For this YAML file, please check the
config-template/sm-config-example.yaml for details.

The YAML configuration is internally converted to the binary format which the
extended IO controller accepts.

Currently only empty slot is avaiable. More modules will be supported later.

Signed-off-by: Baocheng Su <baocheng.su@siemens.com>