iot2050-eio-manager: Add FWU monitoring service
This is useful when user upgrades the software of the AM65x part but
forgets to upgrade the corresponding extended IO controller firmware.

The service periodically reads the current version from eiofs FUSE and
compare with the metadata json. If the firmware need upgrade, it will issue:

-. A `wall` command calling to notify the logged in users.
-. An amber status LED
-. A notification when user uses the iot2050-eio cli to deploy/retrieve
configs.

This is also useful when a same SDCard or USB image is used accorss
different IOT devices which have different version of EIO controller FW.

Signed-off-by: Li Hua Qian <huaqian.li@siemens.com>
Signed-off-by: Baocheng Su <baocheng.su@siemens.com>