Add iot2050 configration web application
Currently the only supported feature is to config the PLC1200 modules.
Web application is easier to use and less error-prone compared to the
YAML configuration, but less efficient for mass deployment.

This also provides a base when new requirement of web application comes.

For the sake of familiarity, react + material UI is used, however in the
future, siemens/ix should be used to get a more pretty and unified UI
style.

This web app should also be possible to pack for the Host PC or server
to be a standalone version, the use case is to manipulate the
configuration but not deploy/retrieve.

Currently this application is only a skeleton, functionalites will be
added later.

Signed-off-by: Baocheng Su <baocheng.su@siemens.com>