refactor: use change-dir to split rootfs parts
This patch refactors the WKS script w.r.t how the rootfs is distributed
across multiple partitions. Previously we used rootfs-dir with a full
rootfs paths plus the sub directory. However, this is not ideal, as the
relation to the rootfs folder is lost and also the new rootfs dir is
technically not a rootfs. While this aparently did not create any issues
in meta-iot2050, we still believe it is semantically incorrect.

This patch switches that over to the WIC option --change-directory that
is provided exactly for this use-case and only deploys a subdirectory
without loosing the rootfs relation with the permissions pseudo dir.

Signed-off-by: Srinuvasan <srinuvasan.a@siemens.com>