iot2050-module-firmware-update: Add module firmware update
The IOT2050 features a customized external signal module, such as the
SM SENS DI. Unlike the standard PLC1200, this module is specifically
tailored to the IOT2050's requirements. As a result, it cannot be
upgraded within the conventional PLC ecosystem. However, we ensure to
provide a method to support firmware upgrades for these custom modules.

And the firmware update for the module should not be included with the
image. This is because the image becomes too large and cumbersome to
release when only the update tool needs to be released. Instead, the
module firmware update should be packaged separately in its own
Debian package for a more efficient and manageable release process.

Signed-off-by: Li Hua Qian <huaqian.li@siemens.com>