board-bootup: setup pwm pins during boot
In order to provide a non-root access to MRAA the PWM pins
have to be configured prior to first access as non-root user.

This is required since a timing issue appears when exporting PWM pins
with MRAA while a potential udev rule is adjusting permissions on the
newly created sysfs entries at the same time.
Subsequent access to a PWM pin may lead to -EPERM until
the permissions have been set properly by udev.

Signed-off-by: Benedikt Niedermayr <benedikt.niedermayr@siemens.com>