mraa: fix non-root access
This solution makes use of the gpio-chardev implementation of MRAA.
Therefore some fixes for setting up the complex pins (multiplexer setup)
had to be done as well as some other minor fixes.

Furthermore a udev rules file that adjusts permissions of all interfaces
is added.

The generation of required users and groups is done in a separate
commit.

Signed-off-by: Benedikt Niedermayr <benedikt.niedermayr@siemens.com>