Add kernel module xt_statistic
When a k3s-agent is installed to the recent image and a hazelcast pod is deployed to the iot-2050-node, ip-tables can not be modified by k3s-agent since XT_STATISTIC Kernel module is missing.
This behavior has only been observed jet in the described setup and can be avoided by deleting and recreating the hazelcast pod.

Signed-off-by: Felix Loesch <49273959+LoeFe@users.noreply.github.com>