iot2050-firmware-update: Correct parameter definition
The '-f', '--force', '-n', '--no-backup', and '-d', '--backup-dir'
options are now part of a mutually exclusive group, ensuring that they
cannot be used together.

Having a default backup dir for -d is not reasonable for force update,
and cites an abnormal case where the force update option would not
work correctly. However, it's still necessary for normal update.

Signed-off-by: Li Hua Qian <huaqian.li@siemens.com>